Add some scripts for adding 'deploy to __' buttons, also some clean up
