+++ bin/configure-beta
@@ -1,18 +1,37 @@
 #!/usr/bin/env ruby
 
-
+# First we require a util file that will install a few utility gems
+# and define some utility methods used by other scirpts.
 require "#{__dir__}/configure-scripts/utils"
+
+# Then we check that we have the right ruby version
 require "#{__dir__}/configure-scripts/check_ruby"
+
+# Next we change instances of "Untilted Application" and `untitled_application`
+# into strings that represent the actual name of the new project being started.
 require "#{__dir__}/configure-scripts/update_configs"
 
+# Now we ask if you want a `Deploy to Heroku` and/or `Deploy to Render` button
+require "#{__dir__}/configure-scripts/deploy_button_heroku"
+require "#{__dir__}/configure-scripts/deploy_button_render"
+
+# Now that we're finished making changes to the new repo we commit them
+require "#{__dir__}/configure-scripts/commit_changes"
+
+# We rename `origin` to `bullet-train` for consistency with documented upgrade
+# procedures. Also to prevent any unintended pushes to the starter repo.
 require "#{__dir__}/configure-scripts/rename_origin"
+
+# Now we connect your local repo to GitHub if you want
 require "#{__dir__}/configure-scripts/setup_github"
 
+# Now we check that you have base-level system dependencies installed.
 require "#{__dir__}/configure-scripts/check_postgres"
 require "#{__dir__}/configure-scripts/check_redis"
 require "#{__dir__}/configure-scripts/check_icu"
 require "#{__dir__}/configure-scripts/check_node"
 
+# And finally we `bundle install` & `yarn install`
 puts "Running `bundle install`.".green
 stream "bundle install"
 
@@ -24,20 +43,3 @@ puts "OK, we're done, but at some point you should edit `config/locales/en/appli
 puts ""
 puts "Next you can run `bin/setup`, then `bin/dev` to spawn a local instance, and then you can navigate to http://localhost:3000 to access your new application.".green
 puts ""
-
-exit
-
-
-# TODO: This was for the "Deploy to Render" button. We should ask explicity
-# about Render instead of assuming that we want it. And it really doesn't
-# have anything to do wit whether or not we setup Github.
-unless skip_github
-  original_repo_link = "https://github.com/bullet-train-co/bullet_train"
-  new_repo_link = ask "What is the link to your repository? We will use this to enable the one-click deploy to Render button for your application."
-  replace_in_file("README.example.md", original_repo_link, new_repo_link, /repo=#{original_repo_link}/)
-end
-
-
-
-
-
